// sources/src/utils/FetchBackend/rest/api/items/dto/update-item.dto.ts
interface ItemCharacteristics {
  dp_characteristicId: number;
  dp_value: string;
}

interface ItemGalery {
  dp_photoUrl: string;
}

export default interface UpdateItemDto {
  dp_name: string;
  dp_model: string;
  dp_cost: number;
  dp_photoUrl: string;
  dp_seoKeywords: string;
  dp_seoDescription: string;
  dp_itemCategoryId: number;
  dp_isHidden: boolean;
  dp_itemCharacteristics: ItemCharacteristics[];
  dp_itemGalery: ItemGalery[];
}